logger: don't actually close the console logger
We nil the internal variable, but we should not close the logger itself,
since this leads to STDERR being closed and the interpreter reopening it
for its own needs, or something like that, which ultimately leads to
the interpreter's warning appearing in console (probably due to ruby
itself reopening stderr to be able to output the messages).

If we simply nil the variable, the interpreter is content with sending
us the messages per our shenanigangs, and everything works just as
expected.